+++ fCraft/MapConversion/MapMCSharp.cs
@@ -84,9 +84,9 @@ namespace fCraft.MapConversion {
 
             // Read in the spawn location
             map.Spawn = new Position {
-                X = (short)( bs.ReadInt16() * 32 ),
-                Z = (short)( bs.ReadInt16() * 32 ),
-                Y = (short)( bs.ReadInt16() * 32 ),
+                X = bs.ReadInt16() * 32,
+                Z = bs.ReadInt16() * 32,
+                Y = bs.ReadInt16() * 32,
                 R = bs.ReadByte(),
                 L = bs.ReadByte(),
             };
